requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
kafka-python==2.0.2
watchdog==6.0.0